requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.7
